refactor: reorg ShipProximityHider.lua to Ship/
