+++ TTS_xwing/src/Global.lua
@@ -109,7 +109,7 @@ local ConfigManager = require("TTS_xwing.src.Player.ConfigManager")
 HotkeyModule = require("TTS_xwing.src.Player.HotkeyManager")
 PlayerEventModule = require("TTS_xwing.src.Player.PlayerActionModule")
 DialManagerModule = require("TTS_xwing.src.Dial.DialManager")
-ShipProximityHider = require("TTS_xwing.src.ShipProximityHider")
+ShipProximityHider = require("TTS_xwing.src.Ship.ShipProximityHider")
 StreamManagerModule = require("TTS_xwing.src.Player.StreamManager")
 --StreamManagerModule.SetUI(self)
 --StreamManagerModule.AddPresentation("Purple", "left")
