+++ TTS_xwing/src/Ship/ShipProximityHider.lua
